Better timetables and hfp data for stop markers WIP
